requirements.txt
python-dotenv
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.123.0
requests==2.31.0
